% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linear_model.R
\docType{data}
\name{linear_model}
\alias{linear_model}
\title{Linear Model for Program Week Data}
\format{An object of class "lm".
\describe{
  \item{coef}{coefficient}
  \item{p}{p-value}
  ...
}}
\source{
\url{https://www.preventnow.com/}
}
\usage{
linear_model
}
\description{
A linear regression model trained on age, bmi, and gender features
in week_data to predict week 16 weight loss.
}
\keyword{datasets}

