% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_weight_loss.R
\name{predict_weight_loss}
\alias{predict_weight_loss}
\title{Predict Week 16 Weight Loss}
\usage{
predict_weight_loss(input)
}
\arguments{
\item{input}{data passed on as \code{newdata} to \code{\link{predict}}}
}
\description{
Simple model with three predictors \code{age}, \code{gender}, and \code{bmi}.
}
\examples{
mydata <- data.frame(
   age=c(24, 54, 32, 75),
   gender=c("m", "f", "f", "m"),
   bmi=c(30, 35, 40, 41)
)
predict_weight_loss(mydata)
}

